Unify upgrade tests
