+++ core/lib/multivm/src/versions/vm_fast/tests/mod.rs
@@ -35,11 +35,11 @@ mod simple_execution;
 mod storage;
 mod tracing_execution_error;
 mod transfer;
+mod upgrade;
 /*
 // mod call_tracer; FIXME: requires tracers
 // mod prestate_tracer; FIXME: is pre-state tracer still relevant?
 mod rollbacks;
-mod upgrade;
 mod utils;
 */
